feat(unifiedio): add unifiedio component
