+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ['formats/*', "servers/*", 'app', 'utils', "streamcenter"]
+members = ['formats/*', "servers/*", "unifiedio", 'app', 'utils', "streamcenter"]
 resolver = "3"
