+++ app/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-tokio = { version = "1.42.0", features = ["full"] }
+tokio = { version = "1.44.1", features = ["full"] }
 rtmp-server = { path = "../servers/rtmp" }
 http-server = { path = "../servers/http" }
 rocket = { version = "0.5.1" }
