+++ config/config.example.ini
@@ -1,5 +1,5 @@
 [logger]
-level = trace
+level = debug
 dir = ./logs/
 
 [rtmp_server]
