+++ formats/flv/Cargo.toml
@@ -6,5 +6,5 @@ edition = "2024"
 [dependencies]
 byteorder = "1.5.0"
 thiserror = "2.0.7"
-tokio-util = { version = "0.7.13", features = ["full"] }
+tokio-util = { version = "0.7.14", features = ["full"] }
 amf = { path = "../amf" }
