+++ formats/rtmp/Cargo.toml
@@ -9,7 +9,7 @@ amf = { path = "../amf" }
 byteorder = "1.5.0"
 hmac = "0.12.1"
 sha2 = "0.10.8"
-tokio = { version = "1.42.0", features = ["full"] }
-tokio-util = { version = "0.7.13", features = ["full"] }
+tokio = { version = "1.44.1", features = ["full"] }
+tokio-util = { version = "0.7.14", features = ["full"] }
 utils = { path = "../../utils" }
-log = "0.4.26"
+tracing = "0.1.41"
