+++ servers/rtmp/Cargo.toml
@@ -9,15 +9,16 @@ amf = { path = "../../formats/amf" }
 utils = { path = "../../utils" }
 flv = { path = "../../formats/flv" }
 thiserror = "2.0.7"
-tokio = { version = "1.42.0", features = ["full"] }
-tokio-util = { version = "0.7.13", features = ["full"] }
+tokio = { version = "1.44.1", features = ["full"] }
+tokio-util = { version = "0.7.14", features = ["full"] }
 stream-center = { path = "../../streamcenter" }
+unified-io = { path = "../../unifiedio" }
 
 serde_json = "1.0.133"
 serde = { version = "1.0.216", features = ["derive"] }
 scope-guard = "1.2.0"
 url = "2.5.4"
-log = "0.4.26"
+tracing = "0.1.41"
 [dependencies.uuid]
 version = "1.11.0"
 features = [
