+++ streamcenter/Cargo.toml
@@ -11,9 +11,9 @@ dashmap = "6.1.0"
 lazy_static = "1.5.0"
 serde_json = "1.0.133"
 thiserror = "2.0.7"
-tokio = "1.42.0"
-tokio-util = "0.7.13"
-log = "0.4.26"
+tokio = "1.44.1"
+tokio-util = "0.7.14"
+tracing = "0.1.41"
 [dependencies.uuid]
 version = "1.11.0"
 features = [
