+++ unifiedio/Cargo.toml
@@ -1,11 +1,9 @@
 [package]
-name = "amf"
+name = "unified-io"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-byteorder = "1.5.0"
-either = "1.13.0"
 thiserror = "2.0.7"
-tokio = { version = "1.44.1", features = ["full"] }
+tokio = { version = "1.44.1", features = ["net", "time"] }
 tokio-util = { version = "0.7.14", features = ["full"] }
